Removed all scp things
This functionality is better suited for the PostCmd stuff.
You configure a script that does this for you, this application doesn't need to know about scp.
It will just run your script on completion. You can do anything in that script.
Just make sure you pass the filename as an argument to your script.